Fix potential compile time problem

+++ 6502/cpu.h
@@ -74,7 +74,7 @@ private:
 	memory&          m_memory;         // CPU can get right at the memory
 	uint8_t          m_status_register;
 
-	addressing_function m_addressing_functions[addressing_mode::NUM_ADDRESSING_MODES];
+	addressing_function m_addressing_functions[static_cast<uint8_t>(addressing_mode::NUM_ADDRESSING_MODES)];
 
 	void set_flag(register_bit bit, uint8_t val) { m_status_register = m_status_register & ~(1<<static_cast<uint8_t>(bit)) | (!!val<<static_cast<uint8_t>(bit)); }
 	uint8_t get_flag(register_bit bit) { return (m_status_register >> static_cast<uint8_t>(bit)) & 0x1; }
